+++ Enter The Dead Zone Client/Assets/Scripts/DZEngine/Entities/Components/DistanceJoint.cs
@@ -50,7 +50,7 @@ namespace DeadZoneEngine.Entities.Components
         {
             
         }
-        public DistanceJoint(ulong ID) : base(ID)
+        public DistanceJoint(ushort ID) : base(ID)
         {
 
         }
