+++ Enter The Dead Zone Client/Assets/Scripts/DZEngine/Entities/PhysicalJoint.cs
@@ -10,7 +10,7 @@ namespace DeadZoneEngine.Entities
     {
         public PhysicalJoint() { }
 
-        public PhysicalJoint(ulong ID) : base(ID) { }
+        public PhysicalJoint(ushort ID) : base(ID) { }
 
         public virtual void PreUpdate() { }
         public virtual void IteratedUpdate() { }
